# TESTING: Test experimental patch for fast YAML tags