This is quick and dirty and hardcoded for the tags taxonomy.

```
benchmark                               old ns/op     new ns/op     delta
BenchmarkFrontmatterTags/YAML:1-4       9578          9720          +1.48%
BenchmarkFrontmatterTags/YAML:21-4      33058         33498         +1.33%
BenchmarkFrontmatterTags/YAML:41-4      55317         56242         +1.67%
BenchmarkFrontmatterTags/YAML2:1-4      9430          6427          -31.85%
BenchmarkFrontmatterTags/YAML2:21-4     30950         7425          -76.01%
BenchmarkFrontmatterTags/YAML2:41-4     52169         8637          -83.44%

benchmark                               old allocs     new allocs     delta
BenchmarkFrontmatterTags/YAML:1-4       66             66             +0.00%
BenchmarkFrontmatterTags/YAML:21-4      211            211            +0.00%
BenchmarkFrontmatterTags/YAML:41-4      352            352            +0.00%
BenchmarkFrontmatterTags/YAML2:1-4      67             44             -34.33%
BenchmarkFrontmatterTags/YAML2:21-4     192            44             -77.08%
BenchmarkFrontmatterTags/YAML2:41-4     313            44             -85.94%

benchmark                               old bytes     new bytes     delta
BenchmarkFrontmatterTags/YAML:1-4       5928          5928          +0.00%
BenchmarkFrontmatterTags/YAML:21-4      10920         10920         +0.00%
BenchmarkFrontmatterTags/YAML:41-4      15944         15944         +0.00%
BenchmarkFrontmatterTags/YAML2:1-4      6040          5224          -13.51%
BenchmarkFrontmatterTags/YAML2:21-4     10880         5224          -51.99%
BenchmarkFrontmatterTags/YAML2:41-4     15744         5224          -66.82%
```